pyproject.toml
[project]
name = "acto"
description = "Push-Button End-to-End Testing of Kubernetes Operators/Controllers"
requires-python = ">= 3.10"
version = "0.1.0"
readme = "README.md"
license = { file = "LICENSE" }
dependencies = [
    "deepdiff~=6.3.0",
    "kubernetes==26.1.0",
    "exrex~=0.11.0",
    "jsonschema~=4.17.3",
    "jsonpatch~=1.33",
    "pandas~=2.0.2",
    "PyYAML~=6.0",
    "requests~=2.31.0",
    "pydantic~=2.5.2",
]

[project.optional-dependencies]
dev = [
    "docker~=6.1.3",
    "prometheus_client",
    "pytest~=7.4.0",
    "pytest-cov~=4.1.0",
    "tabulate~=0.9.0",
    "pip-tools~=7.3.0",
    "pre-commit~=3.6.0",
    "ansible-core~=2.16.2",
    "isort",
    "mypy",
    "black",
    "pylint",
    "jsonref",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["acto"]

[tool.black]
line-length = 80
target-version = ["py310", "py311"]

[tool.isort]
profile = "black"
line_length = 80

[tool.coverage.run]
relative_files = true

[tool.pytest.ini_options]
markers = [
    "local: mark a test to run on a local machine",
    "single_bug_reproduction: mark a test to reproduce a single random bug",
    "all_bug_reproduction: mark a test to reproduce all bugs in the suite",
    "kubernetes_engine: mark a test for cluster set up",
]

[tool.pylint."messages control"]
disable = [
    "raw-checker-failed",
    "bad-inline-option",
    "locally-disabled",
    "file-ignored",
    "suppressed-message",
    "useless-suppression",
    "deprecated-pragma",
    "use-implicit-booleaness-not-comparison-to-string",
    "use-implicit-booleaness-not-comparison-to-zero",
    "use-symbolic-message-instead",
    "too-many-return-statements",
    "too-many-branches",
    "too-many-instance-attributes",
    "too-few-public-methods",
    "consider-using-enumerate",
    "too-many-lines",
    "too-many-arguments",
    "too-many-locals",
    "no-else-return",
    "too-many-nested-blocks",
    "too-many-statements",
    "dangerous-default-value",
    "duplicate-code",
    "missing-module-docstring",
]

[tool.mypy]
ignore_missing_imports = true
